python/showwhy-backend/pyproject.toml
[tool.poetry]
name = "showwhy-backend"
version = "0.0.1"
description = ""
authors = ["Andres Morales Esquivel <andresmor@microsoft.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
azure-functions = "^1.9.0"
azure-functions-durable = "^1.1.3"
azure-storage-blob = "^12.9.0"
azure-cosmos = "^4.2.0"
pymongo = "^4.0.1"
dnspython = "^2.2.0"
showwhy-inference = {path="../showwhy-inference/"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
isort = "^5.10.1"
flake8 = "^4.0.1"
black = "^21.12b0"
mypy = "^0.931"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
py_version = 36
profile = "black"
force_single_line = false
combine_as_imports = true
lines_between_types = 1
lines_after_imports = 2
src_paths = ["showwhy_backend", "tests"]
extend_skip = ["setup.py"]

[flake8]
max-line-length = 88
extend-ignore = "E203"

[tool.black]
target-version = ['py36']
force-exclude = '''
.*/setup\.py$
'''


[tool.mypy]
check_untyped_defs = true
ignore_missing_imports = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true
files = "data_wrangling_components"

[tool.coverage.run]
omit = [
    "tests/*",
    ".*",
    "data/*"
    ]